combine summary and detail view tests into ProgressTableContainer tests
